==> Notebooks/01_exploracao_dataset.ipynb <==
# 🐦 Notebook 01 – Exploração do Dataset Xeno-Canto
Este notebook analisa o dataset utilizado no projeto MelodIA, incluindo visualização,
estatísticas de distribuição, amostras de áudio e espectrogramas.

import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import librosa, librosa.display
import numpy as np

metadata = pd.read_csv('data/xenocanto_meta.csv')
metadata.head()

metadata['cnt_species'] = metadata['gen'] + ' ' + metadata['sp']
plt.figure(figsize=(12,6))
metadata['cnt_species'].value_counts().head(20).plot(kind='bar')
plt.title('Top 20 espécies mais frequentes')
plt.show()

## 🎧 Visualização de espectrograma de um áudio de exemplo

audio_path = 'samples/ave1.wav'
y, sr = librosa.load(audio_path)
plt.figure(figsize=(12,5))
S = librosa.feature.melspectrogram(y=y, sr=sr)
librosa.display.specshow(librosa.power_to_db(S, ref=np.max), sr=sr, x_axis='time', y_axis='mel')
plt.colorbar()
plt.show()
